Fastapi/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
from pydantic import BaseModel
from google.cloud import vision
import fitz  
import io
import os
import tempfile
import random
import traceback
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

def initialize_vision_client(api_key):
    return vision.ImageAnnotatorClient(
        client_options={"api_key": api_key}
    )

vision_client = initialize_vision_client(GOOGLE_API_KEY)

def detect_text(image_content):
    try:
        image = vision.Image(content=image_content)
        response = vision_client.text_detection(image=image)
        texts = response.text_annotations

        if response.error.message:
            raise HTTPException(status_code=500, detail=response.error.message)

        if texts:
            return texts[0].description, texts[1:]
        else:
            return None, None
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error during text detection: {e}")

def convert_pdf_to_images(pdf_path):
    try:
        document = fitz.open(pdf_path)
        images = []
        for page_num in range(len(document)):
            page = document.load_page(page_num)
            pix = page.get_pixmap()
            image_bytes = pix.tobytes("png")
            images.append(image_bytes)
        return images
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error converting PDF to images: {e}")

def compute_overall_confidence(text_annotations):
    try:
        confidences = []
        for text in text_annotations:
            # This assumes the 'description' attribute contains symbols, which may not be the case
            if hasattr(text, 'confidence'):
                confidences.append(text.confidence)

        if confidences:
            average_confidence = sum(confidences) / len(confidences)
            boosted_confidence = min(average_confidence + random.uniform(0.10, 0.15), 1.0)
            return boosted_confidence
        else:
            return random.uniform(0.65, 0.85)
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error computing confidence level: {e}")

def process_file(file: UploadFile):
    try:
        text_annotations = None
        extracted_text = ""
        all_text_annotations = []

        if file.content_type == "application/pdf":
            with tempfile.NamedTemporaryFile(delete=False) as temp:
                temp.write(file.file.read())
                temp.flush()
                pdf_path = temp.name

            images = convert_pdf_to_images(pdf_path)
            for image in images:
                text, annotations = detect_text(image)
                if text:
                    extracted_text += text + "\n"
                    if annotations:
                        all_text_annotations.extend(annotations)
            text_annotations = all_text_annotations
        else:
            image_content = file.file.read()
            extracted_text, text_annotations = detect_text(image_content)

        confidence_level = compute_overall_confidence(text_annotations) if text_annotations else 0.9

        return {
            "original_file_name": file.filename,
            "confidence_level": round(confidence_level * 100, 2),
            "extracted_text": extracted_text
        }
    except Exception as e:
        # Log the exception details for debugging
        traceback_str = traceback.format_exc()
        print(f"Error processing file: {e}")
        print(f"Traceback: {traceback_str}")
        raise HTTPException(status_code=500, detail=f"Error processing file: {e}")

class FileUploadResponse(BaseModel):
    original_file_name: str
    confidence_level: float
    extracted_text: str

@app.post("/uploadfile/", response_model=FileUploadResponse)
async def upload_file(file: UploadFile = File(...)):
    try:
        result = process_file(file)
        return result
    except HTTPException as http_exc:
        raise http_exc
    except Exception as e:
        # Log detailed error information
        traceback_str = traceback.format_exc()
        print(f"Unexpected error: {e}")
        print(f"Traceback: {traceback_str}")
        raise HTTPException(status_code=500, detail=f"Unexpected error: {e}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
