Fastapi/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
black==22.12.0
blinker==1.8.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cvzone==1.6.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.85.2
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
flake8==5.0.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-vision==3.7.3
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
imageio==2.34.2
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langsmith==0.1.81
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy==0.982
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
onnxruntime==1.18.1
openai==1.35.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pooch==1.8.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.9.1
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.18.4
pydeck==0.9.1
pyflakes==2.5.0
Pygments==2.18.0
PyMatting==1.1.12
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.5
python-pptx==0.6.23
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.7.24
rembg==2.0.57
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.20.4
streamlit==1.36.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
tifffile==2024.7.2
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.0
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
trove-classifiers==2024.7.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.19.0
virtualenv==20.26.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
